Move business intelligence controller actions, routes and tests to business controller
